% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitter_data.R
\docType{data}
\name{hitter_data}
\alias{hitter_data}
\title{hitter_data}
\format{
A data frame consisting of 2355 observations with 5 variables.

\describe{
\item{AutoPitchType}{Pitch type as determined by TrackMan}
\item{HitType}{Flyball, Popup, Line Drive, or Ground ball}
\item{ExitSpeed}{Initial velocity of the ball off the bat in mph}
\item{ExitAngle}{Intial launch angle of the ball off the bat in degrees}
\item{Result}{Out, Single, Double, Triple, or Home Run}
}
}
\source{
Duke Baseball coaching staff
}
\usage{
hitter_data
}
\description{
TrackMan metrics corresponding to the balls put into play for which an exit velocity was recorded for 60 games during the 2018 college baseball season.
}
\examples{
summary(hitter_data)
}
\keyword{datasets}
