% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitcher_exitvelo_predictions.R
\name{pitcher_exitvelo_predictions}
\alias{pitcher_exitvelo_predictions}
\title{Pitcher Exit Velocity Predictions}
\arguments{
\item{pitchtype}{The type of pitch thrown (fastball, slider, etc.)}

\item{relspeed}{The velocity of the pitch when it leaves the pitcher's hand in mph}

\item{spinrate}{The rate at which the ball is spinning in rpm}

\item{tilt}{The axis on which the ball spins, measured like hours and minutes on a clock, rounded to the nearest 15 minutes}

\item{inducedvertbreak}{The vertical difference in inches between where the ball is at home plate compared to where it would have been had it been affected by gravity alone}

\item{data}{A data frame corresponding to the data set}
}
\value{
Returns a tibble with the predicted exited velocity
}
\description{
Predicts the exit velocity of a ball in play off of a pitch based on four key pitch design metrics.
}
