% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilt_function.R
\name{tilt_function}
\alias{tilt_function}
\title{Tilt Function}
\arguments{
\item{tilt}{The axis on which the ball spins, measured like hours and minutes on a clock, rounded to the nearest 15 minutes}
}
\value{
Returns a length three character vector of the inputted tilt and the two other closest tilts
}
\description{
Provides the given tilt along with the two neighboring tilts
}
