% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitter_predictions.R
\name{hitter_predictions}
\alias{hitter_predictions}
\title{hitter_predictions}
\arguments{
\item{PitchType}{Type of pitch thrown (fastball, slider, etc.)}

\item{HitType}{Description of ball in play (groundball, line drive, flyball, popup)}

\item{ExitSpeed}{The exit velocity in mph}

\item{ExitAngle}{How steeply up or down the ball leaves the bat in degrees}

\item{data}{A data frame corresponding to the training set}
}
\value{
Returns a tibble with the predicted probability for the result being a single, double, triple, home run, or out
}
\description{
Predicts the result of a ball in play based on the exit velocity, launch angle, hit type, and pitch type.
}
